requirements.txt
django
djangorestframework
grpcio
grpcio-tools
mysqlclient
protobuf
python-dotenv